refactor: rename settings components
