+++ browser-extension/app/components/settings/settings-button.vue
