+++ browser-extension/app/components/settings/settings-color-form-control.vue
@@ -1,4 +1,5 @@
 <template>
+  <!-- eslint-disable-next-line @stylistic/max-len -->
   <div class="flex items-center justify-between gap-1 rounded border border-base-content/50 bg-base-100 px-3 py-1 text-base-content">
     <label
       :for="id"
