+++ browser-extension/app/components/settings/settings-modal.vue
@@ -1,20 +1,19 @@
 <template>
-  <div
-    class="absolute inset-x-8 inset-y-14 -translate-x-3 -translate-y-1 rounded bg-neutral p-2 text-xl"
-  >
+  <!-- eslint-disable-next-line @stylistic/max-len -->
+  <div class="absolute inset-x-8 inset-y-14 -translate-x-3 -translate-y-1 rounded bg-neutral p-2 text-xl">
     <fieldset class="flex justify-between gap-1 bg-base-200 p-2">
       <legend class="rounded bg-base-100 px-1 text-xl text-base-content">
         Pastbl Button Colors
       </legend>
-      <pastbl-settings-color-form-control
+      <settings-color-form-control
         v-model="mainButtonStyles.borderColor"
         :label="`Border`"
       />
-      <pastbl-settings-color-form-control
+      <settings-color-form-control
         v-model="mainButtonStyles.color"
         :label="`Text`"
       />
-      <pastbl-settings-color-form-control
+      <settings-color-form-control
         v-model="mainButtonStyles.backgroundColor"
         :label="`Background`"
       />
@@ -23,5 +22,5 @@
 </template>
 <script setup lang="ts">
 import { mainButtonStyles } from "~/utils/settings.store";
-import PastblSettingsColorFormControl from "~/components/pastbl-settings-color-form-control.vue";
+import SettingsColorFormControl from "~/components/settings/settings-color-form-control.vue";
 </script>
