+++ browser-extension/app/components/settings/with-settings-tooltip.vue
